PostCFD

Documentation for PostCFD.

PostCFD.plot_1dprofile_conv

PostCFD.plot_1dprofile_prim

PostCFD.plot_2dprofile_prim

PostCFD.plot_grid_rectangular

PostCFD.read_dividedgrid

PostCFD.read_dividedrestart

PostCFD.read_flow_double

PostCFD.read_flow_single

PostCFD.read_grid_dims

PostCFD.read_grid_double

PostCFD.read_grid_single

PostCFD.read_header

PostCFD.read_restart

PostCFD.readfiles

PostCFD.write_flow_single

PostCFD.write_grid_single

PostCFD.write_restart_double

PostCFD.writefiles

PostCFD.plot_1dprofile_conv — Method

plot 1D profile of conservative variables.
input: xyz_data, flow_data
keywords : filename,gridrange

PostCFD.plot_1dprofile_prim — Method

#plot_1dprofile_prim(xyz,flow,[gridrange]; filename)
#plot_1dprofile_conv(xyz,flow,[gridrange]; filename)

plot 1D profile of primitive(conservative) variables.
input: xyz_data, flow_data
keywords : filename,

PostCFD.plot_2dprofile_prim — Method

plot 2D profile of primitive variables. input: griddata, flowdata keywords : filename,gridrange

PostCFD.plot_grid_rectangular — Method

##plotgridrectangular(xyz, plane::AbstractString)

PostCFD.read_dividedgrid — Method

read divided grid files. (eg. grid files in dir of "/Grid/Work.divide/comp/" in HPC Template. ) This function returns xyz_vector=Vector{Array{Float64}(jmax,kmax,lmax,3)}. Arg1 : name of directory that contains the divided grid files. Arg2 : grid rank(number) indicated by range (eg. 1:2:19) or array (e.g. [1,3,5,⋯,19]) where you will read.

PostCFD.read_dividedrestart — Method

read divided restart files. This function returns xyz_vector=Vector{Array{Float64}(jmax,kmax,lmax,3)}. Arg1 : name of directory that contains the divided restart files. Arg2 : flow rank(number) indicated by range (eg. 1:2:19) or array (e.g. [1,3,5,⋯,19]) where you will read.

PostCFD.read_flow_double — Method

read a file with pl3d format written in little-endian & stream. This function returns qall=Array{Float64}(jmax,kmax,lmax,5).

PostCFD.read_flow_single — Method

read a file with pl3d format written in little-endian & stream. This function returns qall=Array{Float32}(jmax,kmax,lmax,5).

PostCFD.read_grid_dims — Method

Read grid size. This program return Tuple:(jmax,kmax,lmax)

PostCFD.read_grid_double — Method

"DOUBE PRECISION GRID" read a file with xyz format written in "little-endian" & "stream". This function returns xyz=Array{Float64}(jmax,kmax,lmax,3).

PostCFD.read_grid_single — Method

"SIGLE PRECISION GRID" read a file with xyz format written in "little-endian" & "stream". This function returns xyz=Array{Float32}(jmax,kmax,lmax,3).

PostCFD.read_header — Method

read a file of "header&data format" written in "little-endian" & "stream". this function returns Tuple(labels::strings, digit:Num)

PostCFD.read_restart — Method

read a file with "reatart" format written in little-endian" & "stream. this function returns qall=Array{Float64}(jmax,kmax,lmax,5).

PostCFD.readfiles — Method

wrapper fucntion for read files related to HPC_TEMPLATE
+readfiles(filename, mode)     :: read files
+readfiles(compdir, mode, ids) :: read diveded grid/restart files 
                                  in directory "compdir"

mode ∈ [grid (=grid_double), 
        grid_single,
        flow_fv (=pl3d. flow_single),
        restart,
        dims (=size),
        header,
        ---------------
[!NOTE!]divided_xxx is assumed to be double precision[!NOTE!]
        divided_grid(=grid),
        divided_restart,
        divided_restart_old,
        divided_flow,
        divided_flow_old, ]

[!Note!]
        divided_flow_slice

PostCFD.write_flow_single — Method

write pl3d file to visualize by field view.

PostCFD.write_grid_single — Method

write grid file for fieldview

function writegridsingle(filename::String,xyz::Array{T}) Arg1: filename Aeg2: grid data

PostCFD.write_restart_double — Method

write restart flow

writerestartdouble(filename::String,q::Array{T},params::Array{T},nc::Int32) where T +Arg1: filename +Arg2: q +Arg3: params (mach , AoA, time) +Arg4: nc

PostCFD.writefiles — Method

function writefiles(filename, data...; mode) is the wrapper fucntion for write files related to HPC_TEMPLATE

mode ∈ [gridfv(=gridsingle), flowfv(=pl3d, =flowsingle), grid_double, restart]